@@ -23,29 +23,29 @@
     "format:fix": "prettier --write src/**"
   },
   "dependencies": {
-    "@docusaurus/core": "2.4.3",
-    "@docusaurus/plugin-client-redirects": "2.4.3",
-    "@docusaurus/preset-classic": "2.4.3",
-    "@easyops-cn/docusaurus-search-local": "0.36.0",
+    "@docusaurus/core": "3.0.1",
+    "@docusaurus/plugin-client-redirects": "3.0.1",
+    "@docusaurus/preset-classic": "3.0.1",
+    "@easyops-cn/docusaurus-search-local": "0.38.1",
     "@lottiefiles/react-lottie-player": "3.5.3",
-    "@openfga/sdk": "^0.2.10",
     "@openfga/frontend-utils": "^0.2.0-beta.2",
+    "@openfga/sdk": "^0.2.10",
     "@openfga/syntax-transformer": "^0.2.0-beta.5",
     "assert-never": "1.2.1",
     "clsx": "2.0.0",
     "path-browserify": "1.0.1",
-    "prism-react-renderer": "2.1.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "prism-react-renderer": "2.3.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-feather": "2.0.10",
     "sanitize-html": "2.11.0",
-    "swagger-ui-react": "5.9.2"
+    "swagger-ui-react": "5.10.3"
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "3.0.0",
+    "@docusaurus/module-type-aliases": "3.0.1",
     "@tsconfig/docusaurus": "2.0.2",
     "@types/prismjs": "1.26.3",
-    "@types/react": "17.0.41",
+    "@types/react": "18.2.39",
     "@types/react-router-dom": "5.3.3",
     "@types/sanitize-html": "2.9.5",
     "@typescript-eslint/eslint-plugin": "6.13.1",
